requirements.txt
torch>=2.0.0
Requests>=2.31.0
accelerate
bitsandbytes
ffmpeg
ffmpeg-python
pyannote.audio>=3.1.0
pyannote.core>=5.0.0
pyannote.database>=5.0.1
pyannote.metrics>=3.2.1
pyannote.pipeline>=3.0.1
torchaudio>=2.0.0
speechbrain
huggingface_hub[cli]
